chore: proof of concept cache swapper data fetching
