+++ src/components/Trade/hooks/useSwapper/useSwapper.ts
@@ -2,35 +2,28 @@ import { useToast } from '@chakra-ui/react'
 import { Asset } from '@shapeshiftoss/asset-service'
 import {
   avalancheAssetId,
-  avalancheChainId,
   CHAIN_NAMESPACE,
   ChainId,
   cosmosAssetId,
   ethAssetId,
-  ethChainId,
   fromAssetId,
   osmosisAssetId,
   toAccountId,
 } from '@shapeshiftoss/caip'
-import { avalanche, ChainAdapter, ethereum, EvmChainId } from '@shapeshiftoss/chain-adapters'
+import { ChainAdapter, EvmChainId } from '@shapeshiftoss/chain-adapters'
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
 import {
-  CowSwapper,
-  OsmosisSwapper,
   SwapError,
   SwapErrorTypes,
   Swapper,
   SwapperManager,
-  ThorchainSwapper,
   Trade,
   TradeQuote,
   TradeResult,
   TradeTxs,
   UtxoSupportedChainIds,
-  ZrxSwapper,
 } from '@shapeshiftoss/swapper'
 import { KnownChainIds } from '@shapeshiftoss/types'
-import { getConfig } from 'config'
 import debounce from 'lodash/debounce'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { useFormContext, useWatch } from 'react-hook-form'
@@ -43,18 +36,19 @@ import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { BigNumber, bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit, toBaseUnit } from 'lib/math'
-import { getWeb3InstanceByChainId } from 'lib/web3-instance'
 import { AccountSpecifierMap } from 'state/slices/accountSpecifiersSlice/accountSpecifiersSlice'
 import { accountIdToUtxoParams } from 'state/slices/portfolioSlice/utils'
 import {
   selectAccountSpecifiers,
   selectAssetIds,
+  selectFeatureFlags,
   selectFeeAssetById,
   selectPortfolioCryptoBalanceByAssetId,
 } from 'state/slices/selectors'
-import { store, useAppSelector } from 'state/store'
+import { useAppSelector } from 'state/store'
 
 import { calculateAmounts } from './calculateAmounts'
+import { getSwapperManager } from './swapperManager'
 
 const debounceTime = 1000
 
@@ -82,82 +76,6 @@ type DebouncedQuoteInput = {
   selectedCurrencyToUsdRate: BigNumber
 }
 
-// singleton - do not export me, use getSwapperManager
-let _swapperManager: SwapperManager | null = null
-// singleton - do not export me
-// Used to short circuit calls to getSwapperManager if flags have not changed
-let previousFlags: string = ''
-
-const getSwapperManager = async (): Promise<SwapperManager> => {
-  const flags = store.getState().preferences.featureFlags
-  const flagsChanged = previousFlags !== JSON.stringify(flags)
-  if (_swapperManager && !flagsChanged) return _swapperManager
-  previousFlags = JSON.stringify(flags)
-
-  // instantiate if it doesn't already exist
-  _swapperManager = new SwapperManager()
-
-  const adapterManager = getChainAdapterManager()
-  const ethWeb3 = getWeb3InstanceByChainId(ethChainId)
-  const avaxWeb3 = getWeb3InstanceByChainId(avalancheChainId)
-
-  /** NOTE - ordering here defines the priority - until logic is implemented in getBestSwapper */
-
-  if (flags.Thor) {
-    await (async () => {
-      const midgardUrl = getConfig().REACT_APP_MIDGARD_URL
-      const thorSwapper = new ThorchainSwapper({
-        midgardUrl,
-        adapterManager,
-        web3: ethWeb3,
-      })
-      await thorSwapper.initialize()
-      _swapperManager.addSwapper(thorSwapper)
-    })()
-  }
-
-  const ethereumChainAdapter = adapterManager.get(
-    KnownChainIds.EthereumMainnet,
-  ) as unknown as ethereum.ChainAdapter
-
-  if (flags.CowSwap) {
-    const cowSwapper = new CowSwapper({
-      adapter: ethereumChainAdapter,
-      apiUrl: getConfig().REACT_APP_COWSWAP_HTTP_URL,
-      web3: ethWeb3,
-    })
-
-    _swapperManager.addSwapper(cowSwapper)
-  }
-
-  const zrxEthereumSwapper = new ZrxSwapper({
-    web3: ethWeb3,
-    adapter: ethereumChainAdapter,
-  })
-  _swapperManager.addSwapper(zrxEthereumSwapper)
-
-  if (flags.Avalanche) {
-    const avalancheChainAdapter = adapterManager.get(
-      KnownChainIds.AvalancheMainnet,
-    ) as unknown as avalanche.ChainAdapter
-
-    const zrxAvalancheSwapper = new ZrxSwapper({
-      web3: avaxWeb3,
-      adapter: avalancheChainAdapter,
-    })
-    _swapperManager.addSwapper(zrxAvalancheSwapper)
-  }
-
-  if (flags.Osmosis) {
-    const osmoUrl = getConfig().REACT_APP_OSMOSIS_NODE_URL
-    const cosmosUrl = getConfig().REACT_APP_COSMOS_NODE_URL
-    const osmoSwapper = new OsmosisSwapper({ adapterManager, osmoUrl, cosmosUrl })
-    _swapperManager.addSwapper(osmoSwapper)
-  }
-
-  return _swapperManager
-}
-
 export const useSwapper = () => {
   const toast = useToast()
   const translate = useTranslate()
@@ -175,11 +93,13 @@ export const useSwapper = () => {
   // Swappers will be added in the useEffect below
   const [swapperManager, setSwapperManager] = useState<SwapperManager>(() => new SwapperManager())
 
+  const flags = useSelector(selectFeatureFlags)
+
   useEffect(() => {
     ;(async () => {
-      setSwapperManager(await getSwapperManager())
+      setSwapperManager(await getSwapperManager(flags))
     })()
-  }, [])
+  }, [flags])
 
   const {
     state: { wallet },
